build(tam): disable spell-check for release notes
In the release notes we reference the "Contact Informations" SMT. In
english the plural of information is wrong and `mkdocs-spellcheck`
rightfully marks it as an error. However, at the moment, there is no way
to skip spellchecking for regions only (see
https://github.com/pawamoy/mkdocs-spellcheck/issues/33 for details).

For now we disable spell checking for the whole file.